refacekit-local/ops/apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
redis==5.0.7
python-dotenv==1.0.1
pandas==2.2.2
python-multipart==0.0.9
